% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputoutput.R
\name{conf.file.to.list}
\alias{conf.file.to.list}
\title{Sampled configurations as R lists.}
\usage{
conf.file.to.list(filename, burn.in = 0)
}
\arguments{
\item{filename}{the file  where the sampled configurations are saved.}

\item{burn.in}{the number of  sampled configurations  that are omitted (i.e., the first \code{burn.in} configurations are not loaded).
If configurations are written out  at every iteration, this is the standard burn-in.}
}
\value{
The output is a list whose elements are lists. The i-th list, list[[\link{i}]], represents a configuration, whose elements are
its components:  list[[\link{i}]][[\link{j}]]$x is the vector of the indices of the predictors and  list[[\link{i}]][[\link{j}]]$y that of the responses,
in the j-th component of the i-th configuration.
}
\description{
\code{conf.file.to.list} is used to load into R the sampled configurations saved by the function \code{\link{spavs}} in a file.
}
\examples{
  conf.file.to.list("run1_t1_conf.txt")
}
