% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputoutput.R
\name{pairwise.probs}
\alias{pairwise.probs}
\title{Posterior Pair-wise Probabilities}
\usage{
pairwise.probs(filename, p, q, burn.in = 0)
}
\arguments{
\item{filename}{the file  where the sampled configurations are saved.
The file must be (in the same format as) the file with suffix "_conf.txt" output by  \code{\link{spavs}}.}

\item{p, q}{the number of regressors and responses, respectively.}

\item{burn.in}{the number of  sampled configurations  that are omitted in the computation of the pairwise probabilities
(i.e., the first \code{burn.in} configurations are not used to compute the probabilities).
If configurations are written out  at every iteration, this is the standard burn-in.}
}
\value{
The output is a list with elements:\preformatted{$samples  number of configurations on which the inference is based
          (number of configurations saved in the input file minus burn.in) 

$total_comp total number of components on which the inference is based
           
$pxx the matrix PXX 

$pyy the matrix PYY 

$pxy the matrix PXY
}
}
\description{
\code{pairwise.probs}  returns  estimates of   pairwise posterior probabilities
of any two variables belonging to the same component.
Specifically, it outputs: the matrix PXX, whose \eqn{i,j} element gives the relative incidence of the covariates \eqn{X_i, X_j} occurring in the same component, across components and configurations;
the matrix PYY, whose \eqn{i,j} element  estimates the probability that the responses \eqn{Y_i, Y_j} are in the same component;
the matrix PXY, whose \eqn{i,j} element  estimates the probability that the predictor \eqn{X_i} and response \eqn{Y_j} are in the same component.
}
\examples{
  pairwise.probs("run1_t1_conf.txt", p=12, q=7)
}
